true-self-sim/sample-project/self-sim-game-backend/requirement.txt
fastapi
uvicorn[standard]
PyYAML
passlib[bcrypt]
python-jose[cryptography]
python-multipart
sqlalchemy
asyncpg
psycopg2-binary
alembic
pytest
httpx
aiosqlite
pytest-asyncio